新增用户界面，清除切换页时关闭的错误。优化登录页，优化主页。

+++ .code/ChmlFrpLauncher-cs/StartWindow.xaml.cs
@@ -30,19 +30,17 @@ namespace ChmlFrpLauncher_cs
         {
             Application.Current.Dispatcher.Invoke(() =>
             { 
-                start:
                 //创建路径函数
-                string directoryPath = Directory.GetCurrentDirectory(); string CFL = Path.Combine(directoryPath, "CFL");string frp = Path.Combine(CFL, "frp");string temp = Path.Combine(CFL, "temp");string frpc = Path.Combine(frp, "frpc.exe");string ini = Path.Combine(CFL, "Setup.ini");
+                string directoryPath = Directory.GetCurrentDirectory(); string CFL = Path.Combine(directoryPath, "CFL"); string page_image = Path.Combine(CFL, "pictures"); string frp = Path.Combine(CFL, "frp");string temp = Path.Combine(CFL, "temp");string frpc = Path.Combine(frp, "frpc.exe");string ini = Path.Combine(CFL, "Setup.ini");
                 //创建ini实例
                 var parser = new FileIniDataParser();
                 IniData data;
                 //检测是否有相关配置文件
-                if (!File.Exists(CFL) && !File.Exists(frp) && !File.Exists(ini) && !File.Exists(temp))
+                if (!File.Exists(CFL) && !File.Exists(frp) && !File.Exists(ini) && !File.Exists(page_image) && !File.Exists(temp))
                 {
-                    Directory.CreateDirectory(CFL); Directory.CreateDirectory(frp); Directory.CreateDirectory(temp); //创建文件夹
+                    Directory.CreateDirectory(CFL); Directory.CreateDirectory(frp); Directory.CreateDirectory(page_image); Directory.CreateDirectory(temp); //创建文件夹
                     data = new IniData();
                     data["ChmlFrpLauncher_cs Setup"]["Versions"] = "0.0.0.0.3";
-                    //data["ChmlFrpLauncher_cs Setup"]["number"] = "0";
                     parser.WriteFile(ini, data);
                 }
                 for (int i = 1; i < 6; i++)
@@ -52,12 +50,6 @@ namespace ChmlFrpLauncher_cs
                         File.Create(Path.Combine(CFL, i + ".logs")); //创建logs日志文件
                     }
                 }
-                //检测ini文件创建正确
-                try { data = parser.ReadFile(ini);string V = data["ChmlFrpLauncher_cs Setup"]["Versions"]; }
-                catch{ File.Delete(ini);goto start; }
-                //try{ data = parser.ReadFile(ini);string N = data["ChmlFrpLauncher_cs Setup"]["number"]; }
-                //catch{ File.Delete(ini);goto start; }
-
                 //界面退出，弹出MainWindow。
                 MainWindow MainWindow = new MainWindow();
                 Window window = Window.GetWindow(this);
